[BE] Refactor: Dto 패키지 변경
